fix(test): capture drained the pipe only after the call, which deadlocks
Two failures, both mine, both on the legs I had not run locally.

WINDOWS: TestRunDashboardsImportsAndChecks hung for the full 30-minute
timeout while passing here in a second. cmd/mikroscope's capture helper
redirected os.Stdout into a pipe and read it only after the call returned — a
pipe holds one buffer and blocks its writer when that fills, so any output
larger than the buffer deadlocks, and the buffer's size is the platform's
rather than Go's. `dashboards check` prints a line per panel, which fits the
64 KiB Linux gives and did not fit what Windows does. It drains in a goroutine
now, and TestCaptureSurvivesMoreOutputThanAPipeHolds fails — by timing out —
if that is undone.

LINUX/ARM: a nanosecond wall clock written as an untyped constant does not fit
a 32-bit int. It is int64 now.

`go vet` over ./cmd/... ./internal/... ./test/... passes for all five targets
the cross-vet checks: windows/amd64, darwin/arm64, linux/arm64, linux/arm and
freebsd/amd64. Running that is what would have caught both before the push.

+++ cmd/mikroscope/runrecord_cover_test.go
@@ -34,8 +34,10 @@ func fakeAgentServer(t *testing.T) (subnet string, port int) {
 	var mu sync.Mutex
 	mux := http.NewServeMux()
 	mux.HandleFunc("/healthz", func(w http.ResponseWriter, _ *http.Request) {
+		// int64: a nanosecond wall clock does not fit the 32-bit int of
+		// linux/arm, which is one of the platforms this is type-checked for.
 		fmt.Fprintf(w, `{"ok":true,"seq":%d,"oldest_seq":1,"rate_hz":10,"version":"1.0.9","board":"RB5009","wall_ns":%d}`,
-			100, 1788000000000000000)
+			100, int64(1788000000000000000))
 	})
 	mux.HandleFunc("/snapshot", func(w http.ResponseWriter, r *http.Request) {
 		mu.Lock()
